Add fold and resize functions specific for sparse.coo_matrices.
Hit finders return hits by the coordinates, which is naturally a COO matrix. Sometimes one wants to fold single-shot VMI image, which is much sparser than the sum image. In this case, it is more efficient to fold and resize the sparse matrix rather than the densified.

Unit test showed that the functions for sparse matrices return the same results as those original functions that are designed for dense matrices.